Remove the outdated dist assets
Deleted the dist/assets/index.15e8c738.js and dist/assets/index.72ded636.css files. These were the remaining assets from the old build and are no longer necessary after the recent updated deployment. This cleanup helps to maintain the integrity of the codebase and reduce confusion because these outdated assets are no longer used in the current iteration of the application.